+++ spec/features/alliance_corporations_spec.rb
@@ -4,9 +4,9 @@ require "spec_helper"
 
 describe "List alliance's corporations" do
   context "when etag not set" do
-    let(:options) { {alliance_id: 99_005_443} }
+    let(:options) { {alliance_id: 99_008_595} }
 
-    before { VCR.insert_cassette "esi/alliance_corporations/99005443" }
+    before { VCR.insert_cassette "esi/alliance_corporations/99008595" }
 
     after { VCR.eject_cassette }
 
@@ -16,26 +16,26 @@ describe "List alliance's corporations" do
 
     specify { expect(subject.not_modified?).to eq(false) }
 
-    specify { expect(subject.corporation_ids.size).to eq(21) }
+    specify { expect(subject.corporation_ids.size).to eq(5) }
 
-    specify { expect(subject.corporation_ids.first).to eq(98_192_510) }
+    specify { expect(subject.corporation_ids.first).to eq(98_250_226) }
 
-    specify { expect(subject.etag).to eq("73da1d3efe2beb32798b8cc6a828ada6cd620fca85f78583a1eb509d") }
+    specify { expect(subject.etag).to eq("c8f68430d2ee6e7c342e9730d3ee4bd0aa2d9eb07c6c493c02cdf3b6") }
 
     specify { expect(subject.error_limit_remain).to eq(100) }
 
-    specify { expect(subject.error_limit_reset).to eq(60) }
+    specify { expect(subject.error_limit_reset).to eq(49) }
   end
 
   context "when etag is set" do
     let(:options) do
       {
-        alliance_id: 99_005_443,
-        etag: "73da1d3efe2beb32798b8cc6a828ada6cd620fca85f78583a1eb509d"
+        alliance_id: 99_008_595,
+        etag: "c8f68430d2ee6e7c342e9730d3ee4bd0aa2d9eb07c6c493c02cdf3b6"
       }
     end
 
-    before { VCR.insert_cassette "esi/alliance_corporations/99005443_with_etag" }
+    before { VCR.insert_cassette "esi/alliance_corporations/99008595_with_etag" }
 
     after { VCR.eject_cassette }
 
@@ -43,10 +43,10 @@ describe "List alliance's corporations" do
 
     specify { expect(subject.not_modified?).to eq(true) }
 
-    specify { expect(subject.etag).to eq("73da1d3efe2beb32798b8cc6a828ada6cd620fca85f78583a1eb509d") }
+    specify { expect(subject.etag).to eq("c8f68430d2ee6e7c342e9730d3ee4bd0aa2d9eb07c6c493c02cdf3b6") }
 
     specify { expect(subject.error_limit_remain).to eq(100) }
 
-    specify { expect(subject.error_limit_reset).to eq(20) }
+    specify { expect(subject.error_limit_reset).to eq(10) }
   end
 end