+++ spec/features/alliance_icon_spec.rb
@@ -4,9 +4,9 @@ require "spec_helper"
 
 describe "Get alliance icon" do
   context "when etag not set" do
-    let(:options) { {alliance_id: 99_005_443} }
+    let(:options) { {alliance_id: 99_008_595} }
 
-    before { VCR.insert_cassette "esi/alliance_icon/99005443" }
+    before { VCR.insert_cassette "esi/alliance_icon/99008595" }
 
     after { VCR.eject_cassette }
 
@@ -17,22 +17,22 @@ describe "Get alliance icon" do
     specify { expect(subject.not_modified?).to eq(false) }
 
     specify do
-      expect(subject.as_json).to eq(icon_medium: "https://images.evetech.net/alliances/99005443/logo?tenant=tranquility&size=128",
-        icon_small: "https://images.evetech.net/alliances/99005443/logo?tenant=tranquility&size=64")
+      expect(subject.as_json).to eq(icon_medium: "https://images.evetech.net/alliances/99008595/logo?tenant=tranquility&size=128",
+        icon_small: "https://images.evetech.net/alliances/99008595/logo?tenant=tranquility&size=64")
     end
 
-    specify { expect(subject.etag).to eq("d067edb53a9019e0e05d99a2b1e7c7745433a982dda16667a14bff49") }
+    specify { expect(subject.etag).to eq("1e9455f8ecc240e342f79e9225582c582205c50995d5d4d696b87601") }
 
     specify { expect(subject.error_limit_remain).to eq(100) }
 
-    specify { expect(subject.error_limit_reset).to eq(30) }
+    specify { expect(subject.error_limit_reset).to eq(10) }
   end
 
   context "when etag is set" do
     let(:options) do
       {
-        alliance_id: 99_005_443,
-        etag: "d067edb53a9019e0e05d99a2b1e7c7745433a982dda16667a14bff49"
+        alliance_id: 99_008_595,
+        etag: "1e9455f8ecc240e342f79e9225582c582205c50995d5d4d696b87601"
       }
     end
 
@@ -44,7 +44,7 @@ describe "Get alliance icon" do
 
     specify { expect(subject.not_modified?).to eq(true) }
 
-    specify { expect(subject.etag).to eq("d067edb53a9019e0e05d99a2b1e7c7745433a982dda16667a14bff49") }
+    specify { expect(subject.etag).to eq("1e9455f8ecc240e342f79e9225582c582205c50995d5d4d696b87601") }
 
     specify { expect(subject.error_limit_remain).to eq(100) }