+++ spec/features/alliance_spec.rb
@@ -4,9 +4,9 @@ require "spec_helper"
 
 describe "Get alliance information" do
   context "when etag not set" do
-    let(:options) { {alliance_id: 99_005_443} }
+    let(:options) { {alliance_id: 99_008_595} }
 
-    before { VCR.insert_cassette "esi/alliances/99005443" }
+    before { VCR.insert_cassette "esi/alliances/99008595" }
 
     after { VCR.eject_cassette }
 
@@ -17,31 +17,31 @@ describe "Get alliance information" do
     specify { expect(subject.scope).to eq(nil) }
 
     specify do
-      expect(subject.as_json).to eq(name: "Kids With Guns Alliance",
-        creator_id: 94_195_096,
-        creator_corporation_id: 98_306_624,
-        ticker: "-KWG-",
-        date_founded: "2015-05-03T19:45:17Z",
-        executor_corporation_id: 98_306_624,
+      expect(subject.as_json).to eq(name: "The Dead Parrots",
+        creator_id: 95_078_959,
+        creator_corporation_id: 98_573_850,
+        ticker: "TDP",
+        date_founded: "2018-09-11T19:55:16Z",
+        executor_corporation_id: 98_565_696,
         faction_id: nil)
     end
 
-    specify { expect(subject.etag).to eq("6780e53a01c7d9715b5f445126c4f2c137da4be79e4debe541ce3ab2") }
+    specify { expect(subject.etag).to eq("5eedb79d062aad2bed4e103e149ebf728be9f2d69b63ecf186f45c1d") }
 
     specify { expect(subject.error_limit_remain).to eq(100) }
 
-    specify { expect(subject.error_limit_reset).to eq(29) }
+    specify { expect(subject.error_limit_reset).to eq(1) }
   end
 
   context "when etag is set" do
     let(:options) do
       {
-        alliance_id: 99_005_443,
-        etag: "6780e53a01c7d9715b5f445126c4f2c137da4be79e4debe541ce3ab2"
+        alliance_id: 99_008_595,
+        etag: "5eedb79d062aad2bed4e103e149ebf728be9f2d69b63ecf186f45c1d"
       }
     end
 
-    before { VCR.insert_cassette "esi/alliances/99005443_with_etag" }
+    before { VCR.insert_cassette "esi/alliances/99008595_with_etag" }
 
     after { VCR.eject_cassette }
 
@@ -49,10 +49,10 @@ describe "Get alliance information" do
 
     specify { expect(subject.not_modified?).to eq(true) }
 
-    specify { expect(subject.etag).to eq("6780e53a01c7d9715b5f445126c4f2c137da4be79e4debe541ce3ab2") }
+    specify { expect(subject.etag).to eq("5eedb79d062aad2bed4e103e149ebf728be9f2d69b63ecf186f45c1d") }
 
     specify { expect(subject.error_limit_remain).to eq(100) }
 
-    specify { expect(subject.error_limit_reset).to eq(28) }
+    specify { expect(subject.error_limit_reset).to eq(25) }
   end
 end